rename metadata module to ingest-csv
